fix test_advection and docstring
