+++ model/atmosphere/advection/tests/advection/integration_tests/test_advection.py
@@ -6,15 +6,12 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-import gt4py.next.typing as gtx_typing
 import pytest
 
-import icon4py.model.testing.test_utils as test_helpers
 from icon4py.model.atmosphere.advection import advection
 from icon4py.model.common import constants, dimension as dims
 from icon4py.model.common.decomposition import definitions as decomposition
 from icon4py.model.common.grid import (
-    base as base_grid,
     geometry_attributes as geometry_attrs,
     horizontal as h_grid,
 )
@@ -22,9 +19,7 @@ from icon4py.model.common.interpolation.interpolation_fields import compute_lsq_
 from icon4py.model.common.utils import data_allocation as data_alloc
 from icon4py.model.testing import (
     definitions,
-    grid_utils,
     grid_utils as gridtest_utils,
-    serialbox as sb,
 )
 from icon4py.model.testing.fixtures.datatest import (
     backend,
@@ -127,14 +122,6 @@ def test_advection_run_single_step(  # noqa: PLR0917 [too-many-positional-argume
     advection_exit_savepoint,
     experiment: definitions.Experiment,
 ):
-    # TODO(OngChia): the last datatest fails on GPU (or even CPU) backend when there is no advection because the horizontal flux is not zero. Further check required.
-    if (
-        even_timestep
-        and horizontal_advection_type == advection.HorizontalAdvectionType.NO_ADVECTION
-    ):
-        pytest.xfail(
-            "This test is skipped until the cause of nonzero horizontal advection if revealed."
-        )
     config = advection.AdvectionConfig(
         horizontal_advection_type=horizontal_advection_type,
         horizontal_advection_limiter=horizontal_advection_limiter,
@@ -154,10 +141,9 @@ def test_advection_run_single_step(  # noqa: PLR0917 [too-many-positional-argume
         domain_length=geometry.grid.grid_params.domain_length,
         domain_height=geometry.grid.grid_params.domain_height,
         grid_sphere_radius=constants.EARTH_RADIUS,
-        lsq_dim_unk=2,
-        lsq_dim_c=3,
-        lsq_wgt_exp=2,
-        lsq_dim_stencil=3,
+        lsq_dim_unk=experiment.config.interpolation.lsq_dim_unk,
+        lsq_dim_c=experiment.config.interpolation.lsq_dim_c,
+        lsq_wgt_exp=experiment.config.interpolation.lsq_wgt_exp,
         start_idx=icon_grid.start_index(
             h_grid.domain(dims.CellDim)(h_grid.Zone.LATERAL_BOUNDARY_LEVEL_2)
         ),
@@ -218,69 +204,4 @@ def test_advection_run_single_step(  # noqa: PLR0917 [too-many-positional-argume
         p_tracer_new=p_tracer_new,
         p_tracer_new_ref=p_tracer_new_ref,
         even_timestep=even_timestep,
-    )
-
-
-@pytest.mark.level("unit")
-@pytest.mark.datatest
-def test_compute_lsq_coeffs(
-    icon_grid: base_grid.Grid,
-    grid_savepoint: sb.IconGridSavepoint,
-    backend: gtx_typing.Backend,
-    interpolation_savepoint: sb.InterpolationSavepoint,
-    experiment: definitions.Experiment,
-) -> None:
-    gm = grid_utils.get_grid_manager_from_identifier(
-        experiment.grid,
-        num_levels=1,
-        keep_skip_values=True,
-        allocator=backend,
-    )
-
-    c2e2c = gm.grid.connectivities["C2E2C"].asnumpy()
-    cell_owner_mask = grid_savepoint.c_owner_mask().asnumpy()
-    grid_sphere_radius = constants.EARTH_RADIUS
-    lsq_dim_unk = 2
-    lsq_dim_c = 3
-    lsq_wgt_exp = 2
-    cell_domain = h_grid.domain(dims.CellDim)
-
-    min_rlcell_int = gm.grid.end_index(cell_domain(h_grid.Zone.LOCAL))
-    start_idx = gm.grid.start_index(cell_domain(h_grid.Zone.LATERAL_BOUNDARY_LEVEL_2))
-
-    grid_geometry = grid_utils.get_grid_geometry(backend, experiment.grid, experiment.config)
-    cell_center_x = grid_geometry.get(geometry_attrs.CELL_CENTER_X).asnumpy()
-    cell_center_y = grid_geometry.get(geometry_attrs.CELL_CENTER_Y).asnumpy()
-    domain_length = gm.grid.grid_params.domain_length
-    domain_height = gm.grid.grid_params.domain_height
-    lsq_dim_stencil = 3
-
-    coordinates = gm.coordinates
-    cell_lat = coordinates[dims.CellDim]["lat"].asnumpy()
-    cell_lon = coordinates[dims.CellDim]["lon"].asnumpy()
-    lsq_pseudoinv = compute_lsq_coeffs(
-        cell_center_x=cell_center_x,
-        cell_center_y=cell_center_y,
-        cell_lat=cell_lat,
-        cell_lon=cell_lon,
-        c2e2c=c2e2c,
-        cell_owner_mask=cell_owner_mask,
-        domain_length=domain_length,
-        domain_height=domain_height,
-        grid_sphere_radius=grid_sphere_radius,
-        lsq_dim_unk=lsq_dim_unk,
-        lsq_dim_c=lsq_dim_c,
-        lsq_wgt_exp=lsq_wgt_exp,
-        lsq_dim_stencil=lsq_dim_stencil,
-        start_idx=start_idx,
-        min_rlcell_int=min_rlcell_int,
-        geometry_type=icon_grid.grid_params.geometry_type,
-        exchange=decomposition.single_node_exchange,
-    )
-
-    assert test_helpers.dallclose(
-        interpolation_savepoint.lsq_pseudoinv_1().asnumpy(), lsq_pseudoinv[:, 0, :], atol=1e-15
-    )
-    assert test_helpers.dallclose(
-        interpolation_savepoint.lsq_pseudoinv_2().asnumpy(), lsq_pseudoinv[:, 1, :], atol=1e-15
     )
